backend/requirements.txt
flask
flask-cors
tensorflow
librosa
speechrecognition
requests
python-dotenv
google-cloud-firestore